+++ terminal.h
@@ -21,10 +21,15 @@
 #ifndef UL_TERMINAL_H
 #define UL_TERMINAL_H
 
+#include <stdbool.h>
+
 /**
  * Prepare the current TTY for graphics output.
+ *
+ * @param enable_graphics_mode if true, switch terminal into graphics mode (hides command prompt)
+ * @param disable_keyboard_input if true, disable terminal keyboard input (hides entered text)
  */
-void ul_terminal_prepare_current_terminal(void);
+void ul_terminal_prepare_current_terminal(bool enable_graphics_mode, bool disable_keyboard_input);
 
 /**
  * Reset the current TTY to text output.